rename and cleanup components
